refactor EthStatusDataProvider
